refactor(policy): split runtime shape owners
